Adding new application changes to admin dashboard
