+++ admin_view/src/components/ApplicationInterface.vue
@@ -1,0 +1,5 @@
+<template>
+  <div class="main_section">
+    <h4>Application Interface</h4>
+  </div>
+</template>
