+++ admin_view/src/components/DashboardItem.vue
