+++ admin_view/src/main.js
@@ -4,17 +4,18 @@ import VueRouter from 'vue-router'
 
 import ExperimentsDashboard from './dashboards/ExperimentDashboard.vue'
 import AdminDashboard from './dashboards/AdminDashboard.vue'
+import router from './router'
+Vue.config.productionTip = false;
 
-Vue.config.productionTip = false
-
-Vue.use(VueResource)
-Vue.use(VueRouter)
+Vue.use(VueResource);
+Vue.use(VueRouter);
 
 
 
 export function initializeApacheAiravataDashboard(dashboardName) {
   return new Vue({
     el: '#app',
+    router,
     template: '<' + dashboardName + '/>',
     components: {ExperimentsDashboard, AdminDashboard}
